main/requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
drf-yasg==1.21.7
Faker==24.4.0
inflection==0.5.1
packaging==24.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
